third_party/blink/web_tests/platform/win10/virtual/webnn-service-on-cpu/external/wpt/webnn/conformance_tests/subgraph.https.any.serviceworker_cpu-expected.txt

This is a testharness.js-based test.
[FAIL] batchNormalization options.axis=0 + gelu
  assert_less_than_equal: assert_array_approx_equals_ulp: test batchNormalization gelu float32 actual -0.035951387137174606 should be close enough to expected -0.03595117852091789 by ULP distance: expected a number less than or equal to 24n but got 56n
Harness: the test ran to completion.
